Move and remove shared components
