+++ packages/transaction/detail/info/transactionAmount/index.js
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { MODULE_ASSETS_NAME_ID_MAP } from '@transaction/configuration/moduleAssets';
 import LiskAmount from '@shared/liskAmount';
-import DiscreetMode from '../discreetMode';
+import DiscreetMode from '@shared/discreetMode';
 import styles from './transactionAmount.css';
 
 const getTxDirectionConfig = (moduleAssetId, host, recipient) => {
