+++ packages/views/screens/managers/login/login.js
@@ -10,7 +10,7 @@ import Piwik from '@common/utilities/piwik';
 import { PrimaryButton } from '@basics/buttons';
 import PassphraseInput from '@wallet/manage/passphraseInput';
 import Icon from '@basics/icon/index';
-import DiscreetModeToggle from '@shared/discreetModeToggle';
+import DiscreetModeToggle from '@settings/setters/toggles/discreetModeToggle';
 import NetworkSelector from '../../../../settings/setters/networkSelector';
 import RecoveryPhrase from './recoveryPhrase';
 import styles from './login.css';
