+++ packages/views/screens/managers/send/summary/summary.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { mount } from 'enzyme';
 import { tokenMap } from '@token/configuration/tokens';
 import { toRawLsk } from '@token/utilities/lsk';
-import { formatAmountBasedOnLocale } from '@common/utilities/formattedNumber';
+import { formatAmountBasedOnLocale } from '@views/shared/formattedNumber';
 import i18n from '@setup/i18n/i18n';
 import accounts from '@tests/constants/accounts';
 import Summary from './summary';
