--- packages/views/shared/analyticsDialog/analyticsDialog.css
@@ -1,4 +1,0 @@
-
-.buttons {
-  max-width: 170px !important;
-}
