+++ packages/views/shared/customRoute/index.js
@@ -7,7 +7,7 @@ import Piwik from '@common/utilities/piwik';
 import routes from '@screens/router/routes';
 import Login from '@screens/managers/login';
 import offlineStyle from '@basics/offlineWrapper/offlineWrapper.css';
-import ErrorBoundary from '../errorBoundary';
+import ErrorBoundary from './errorBoundary';
 
 const checkNetwork = state =>
   !!state.network.name
