+++ packages/wallet/detail/info/walletDetails/walletDetails.js
@@ -13,7 +13,7 @@ import Converter from '@shared/converter';
 import { fromRawLsk } from '@token/utilities/lsk';
 import LiskAmount from '@shared/liskAmount';
 import LockedBalanceLink from '@screens/managers/wallet/overview/balanceInfo/unlocking';
-import DiscreetMode from '../discreetMode';
+import DiscreetMode from '@shared/discreetMode';
 import styles from './walletDetails.css';
 
 const WalletDetails = ({
